pcl: always save coroutine stack pointer and size
This is needed so that the main coroutine that executes on a
host's stack and therefore doesn't have the coroutine stack pointer
filled in, actually got it filled. Otherwise asan doesn't know to
what stack we switch, when switching back to main coroutine, and that
results in:

==3966254==WARNING: ASan is ignoring requested __asan_handle_no_return: stack type: default top: 0x000000000000; bottom 0x7ffe3979b000; size: 0xffff8001c6865000 (-140729862696960)
False positive error reports may follow
For details see https://github.com/google/sanitizers/issues/189

Which is fixed by this patch.

Thanks to @ioquatix for the suggestion.

+++ src/base/lib/libpcl/pcl_private.h
@@ -58,7 +58,7 @@ typedef struct s_co_base {
 	struct s_co_base *restarget;
 	struct s_cothread_ctx *ctx_main;
 	void *tmp;
-	int stack_size;
+	size_t stack_size;
 	char *stack;
 	unsigned int exited:1;
 } co_base;
